Move insert_copy_discard into Linearizer
