+++ hugr-passes/src/lower_types.rs
@@ -17,9 +17,12 @@ use hugr_core::ops::{
 use hugr_core::types::{
     CustomType, Signature, Transformable, Type, TypeArg, TypeEnum, TypeTransformer,
 };
-use hugr_core::{Hugr, IncomingPort, Node, OutgoingPort};
+use hugr_core::{Hugr, Node};
 
 use crate::validation::{ValidatePassError, ValidationLevel};
+
+mod linearize;
+pub use linearize::Linearizer;
 
 /// A thing to which an Op can be lowered, i.e. with which a node can be replaced.
 #[derive(Clone, Debug, PartialEq)]
@@ -74,22 +77,7 @@ pub struct LowerTypes {
     type_map: HashMap<CustomType, Type>,
     /// Parametric types are handled by a function which receives the lowered typeargs.
     param_types: HashMap<ParametricType, Arc<dyn Fn(&[TypeArg]) -> Option<Type>>>,
-    // Keyed by lowered type, as only needed when there is an op outputting such
-    copy_discard: HashMap<Type, (OpReplacement, OpReplacement)>,
-    // Copy/discard of parametric types handled by a function that receives the new/lowered type.
-    // We do not allow linearization to "parametrized" non-extension types, at least not
-    // in one step. We could do that using a trait, but it seems enough of a corner case.
-    // Instead that can be achieved by *firstly* lowering to a custom linear type, with copy/dup
-    // inserted; *secondly* by lowering that to the desired non-extension linear type,
-    // including lowering of the copy/dup operations to...whatever.
-    copy_discard_parametric: HashMap<
-        ParametricType,
-        // TODO should pass &LowerTypes, or at least some way to call copy_op / discard_op, to these
-        (
-            Arc<dyn Fn(&[TypeArg]) -> OpReplacement>,
-            Arc<dyn Fn(&[TypeArg]) -> OpReplacement>,
-        ),
-    >,
+    linearize: Linearizer,
     // Handles simple cases Op1 -> Op2.
     op_map: HashMap<OpHashWrapper, OpReplacement>,
     // Called after lowering typeargs; return None to use original OpDef
@@ -175,7 +163,7 @@ impl LowerTypes {
     }
 
     pub fn linearize(&mut self, src: Type, copy: OpReplacement, discard: OpReplacement) {
-        self.copy_discard.insert(src, (copy, discard));
+        self.linearize.register(src, copy, discard)
     }
 
     pub fn linearize_parametric(
@@ -184,8 +172,7 @@ impl LowerTypes {
         copy_fn: Box<dyn Fn(&[TypeArg]) -> OpReplacement>,
         discard_fn: Box<dyn Fn(&[TypeArg]) -> OpReplacement>,
     ) {
-        self.copy_discard_parametric
-            .insert((&src).into(), (Arc::from(copy_fn), Arc::from(discard_fn)));
+        self.linearize.register_parametric(src, copy_fn, discard_fn)
     }
 
     /// Configures this instance to change occurrences of `src` to `dest`.
@@ -301,68 +288,12 @@ impl LowerTypes {
                     if targets.len() != 1 {
                         hugr.disconnect(n, outp);
                         let typ = new_sig.out_port_type(outp).unwrap();
-                        self.insert_copy_discard(hugr, n, outp, typ, &targets)
+                        self.linearize.insert_copy_discard(hugr, n, outp, typ, &targets)
                     }
                 }
             }
         }
         Ok(changed)
-    }
-
-    fn insert_copy_discard(
-        &self,
-        hugr: &mut impl HugrMut,
-        mut src_node: Node,
-        mut src_port: OutgoingPort,
-        typ: &Type,
-        targets: &[(Node, IncomingPort)],
-    ) {
-        let (last_node, last_inport) = match targets.last() {
-            None => {
-                let discard = self
-                    .discard_op(typ)
-                    .expect("Don't know how to discard {typ:?}"); // TODO return error
-
-                let disc = discard.add(hugr, hugr.get_parent(src_node).unwrap());
-                (disc, 0.into())
-            }
-            Some(last) => *last,
-        };
-        if targets.len() > 1 {
-            let copy = self.copy_op(typ).expect("Don't know how copy {typ:?"); // TODO return error
-                                                                               // Could sanity-check signature here?
-            for (tgt_node, tgt_port) in &targets[..targets.len() - 1] {
-                let n = copy.add(hugr, hugr.get_parent(src_node).unwrap());
-                hugr.connect(src_node, src_port, n, 0);
-                hugr.connect(n, 0, *tgt_node, *tgt_port);
-                (src_node, src_port) = (n, 1.into());
-            }
-        }
-        hugr.connect(src_node, src_port, last_node, last_inport);
-    }
-
-    pub fn copy_op(&self, typ: &Type) -> Option<OpReplacement> {
-        if let Some((copy, _)) = self.copy_discard.get(typ) {
-            return Some(copy.clone());
-        }
-        let TypeEnum::Extension(exty) = typ.as_type_enum() else {
-            return None;
-        };
-        self.copy_discard_parametric
-            .get(&exty.into())
-            .map(|(copy_fn, _)| copy_fn(exty.args()))
-    }
-
-    pub fn discard_op(&self, typ: &Type) -> Option<OpReplacement> {
-        if let Some((_, discard)) = self.copy_discard.get(typ) {
-            return Some(discard.clone());
-        }
-        let TypeEnum::Extension(exty) = typ.as_type_enum() else {
-            return None;
-        };
-        self.copy_discard_parametric
-            .get(&exty.into())
-            .map(|(_, discard_fn)| discard_fn(exty.args()))
     }
 
     fn change_node(&self, hugr: &mut impl HugrMut, n: Node) -> Result<bool, ChangeTypeError> {
